backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
motor==3.3.2
pymongo==4.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
bcrypt==4.1.1
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
PyPDF2==3.0.1
reportlab==4.0.7
python-docx==1.1.0
httpx==0.25.2
pydantic==2.5.2
pydantic[email]==2.5.2
python-dateutil==2.8.2
beautifulsoup4
html5lib
pdfplumber
python-docx
weasyprint